requirements.txt
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
cython==3.0.10
fastparquet==2024.5.0
fonttools==4.53.0
fsspec==2024.6.1
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyarrow==16.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
mysql-connector-python==9.2.0
dotenv==0.9.9
torch==2.5.1
darts==0.32.0
